add progress_change test/fix bug with progress count

+++ tests/test_api.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 import pytest
 from fastapi.encoders import jsonable_encoder
 
@@ -6,6 +8,10 @@ from api.books.queries import insert_book, get_books_from_db, get_book_from_db
 from api.books.schemas import BookAdd, BookSchema
 from database import async_session_factory
 from tests.conftest import USER_ID
+
+
+async def form_db_response(data, detail: Optional[str] = None):
+    return {"status": "success", "data": jsonable_encoder(data), "detail": detail}
 
 
 class TestAPI:
@@ -17,9 +23,9 @@ class TestAPI:
         async with async_session_factory() as session:
             result: Book = await insert_book(session, data, user_id)
             book: BookSchema = BookSchema.from_orm(result)
-        response = await get_client.get(f"/book/{book.id}")
+        response = await get_client.get(f"/book/{book.id}?user_id={user_id}")
         assert response.status_code == 200
-        assert response.json() == {"status":"success","data":jsonable_encoder(book),"detail":None}
+        assert response.json() == await form_db_response(book)
 
     @pytest.mark.asyncio
     @pytest.mark.parametrize("user_id", [USER_ID])
@@ -27,7 +33,7 @@ class TestAPI:
         async with async_session_factory() as session:
             result: list[Book] = await get_books_from_db(session, user_id)
             books: list[BookSchema] = list(map(lambda x: BookSchema.from_orm(x), result))
-        response = await get_client.get("/book/list")
+        response = await get_client.get(f"/book/list?user_id={user_id}")
         assert response.status_code == 200
         assert response.json() == {"status":"success","data":jsonable_encoder(books),"detail":None}
 
@@ -36,7 +42,7 @@ class TestAPI:
         (BookAdd(title="Dubrovskiy", author="Pushkin", volume=300), USER_ID)
     ])
     async def test_add_book(self, get_client, data, user_id):
-        response = await get_client.post(f"/book/add?tg_id={user_id}",
+        response = await get_client.post(f"/book/add?user_id={user_id}",
                                          json={"title": data.title,
                                                "author": data.author,
                                                "volume": data.volume})
@@ -45,10 +51,23 @@ class TestAPI:
             result: Book = await get_book_from_db(session, json_data['data']["id"], user_id)
             book: BookSchema = BookSchema.from_orm(result)
         assert response.status_code == 200
-        assert json_data == {"status":"success","data":jsonable_encoder(book),"detail":None}
+        assert json_data == await form_db_response(book)
 
     @pytest.mark.asyncio
-    @pytest.mark.parametrize("book_id", [1, 2])
-    async def test_delete_book(self, get_client, book_id):
-        response = await get_client.delete(f"/book/delete/{book_id}")
+    @pytest.mark.parametrize("book_id, user_id, current_page", [
+        (1, USER_ID, 100), (2, USER_ID, 200)
+    ])
+    async def test_change_progress(self, get_client, book_id, user_id, current_page):
+        response = await get_client.patch(f"/book/progress/{book_id}?user_id={user_id}&current_page={current_page}")
+        json_data = response.json()
+        async with async_session_factory() as session:
+            result: Book = await get_book_from_db(session, json_data['data']["id"], user_id)
+            book: BookSchema = BookSchema.from_orm(result)
+        assert response.status_code == 200
+        assert json_data == await form_db_response(book, 'Книга изменена')
+
+    @pytest.mark.asyncio
+    @pytest.mark.parametrize("book_id, user_id", [(1, USER_ID), (2, USER_ID)])
+    async def test_delete_book(self, get_client, book_id, user_id):
+        response = await get_client.delete(f"/book/delete/{book_id}?user_id={user_id}")
         assert response.status_code == 200
